Rewrite instruction display
After my finding of the SCUMM manual in archive.org, I realized SCUMM
language was not an assembler language, but a high level one. Despite
the authors tried to simplify the bytecode language by adding macro
instructions to it, the game writers actually used a high level
language.

This means it makes no sense to represent the instructions as high level
constructions. This change marks a new direction by making all
instructions to have a short acronym and show argments as if it was an
assembler language.